setting robot heading to a fixed point

+++ 2024 Crescendo/src/main/java/frc/utils/Helpers.java
@@ -22,7 +22,7 @@ public class Helpers
     return input;
   }
 
-  public static int clamp( int val, int min, int max )
+  public static double clamp( double val, double min, double max )
   {
     return Math.max( min, Math.min( max, val ) );
   }
